Index entries by the displayField specified by the content type
In addition to fetching the entries, we now additionally fetch the content_type data. The serializer then uses this information to index each entry by the contents of the specified field as the index.

This does require that the displayField is unique - setting it as a slug ensures this - so it would make sense for this to be an optional feature rather than default. But it allows for precise selection of an entry with requiring looping within your page. Can now use:

{{ site.data.contentful.spaces.mySpace.myContentType.entryName.myField }}